% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdef.R
\name{bdef}
\alias{bdef}
\title{Spatial Deformation via B-spline}
\usage{
bdef(cov.model = RMmatern(nu = 2.5, var = NA, scale = NA) + RMnugget(var =
  NA), x, y, df1 = 8, df2 = 8, window = list(x = range(x[, 1]), y =
  range(x[, 2])), lambda = NA, ...)
}
\arguments{
\item{cov.model}{A model for the covariance function. See \code{\link{RMmodel}}
on the RandomFields package for details. Defaults to
Mat\'{e}rn model with parameters nu = 2.5 and
variance and scale to be estimated.}

\item{x}{A n times 2 matrix of coordinates for the sampled data points.}

\item{y}{A vector of length n with the values taken by the response
at the corresponding spatial locations in x.}

\item{df1}{Square root of the total degrees of freedom for the tensor
product of B-spline basis for the first coordinate.
Defaults to 8.}

\item{df2}{Square root of the total degrees of freedom for the tensor
product of B-spline basis for the second coordinate.
Defaults to 8.}

\item{window}{A list with two named entries, "x" and "y", each a length 2
                   numeric vector with the minimum and maximum values
                   taken for the respective coordinate in the sampled
                   region. Defaults to the range of the columns in
                   argument x.
@param lambda Penalization parameter. If set to NA (default), this
                   parameter will be estimated. Otherwise, it will use
                   the user supplied value. The parameter is in a log
                   scale, thus ranging from -Inf to +Inf.
@param ... Additional arguments for RFfit.}

\item{def.x}{A n times 2 matrix of deformed coordinates,}
}
\value{
\code{bdef} returns an object of \code{\link{class}} "bdef"
containing at least the following components
  \item{window}{The w}
  \item{basis}{list of B-spline basis functions, inherited from
               \code{\link{bs}}, with one set for each coordinate}
  \item{theta1}{B-spline coefficients for the coordinate x1}
  \item{theta2}{B-spline coefficients for the coordinate x2}
}
\description{
This function finds a spatial deformation for Gaussian processes, based
on the penalized log-likelihood function. The deformation is obtained
via the tensor product of B-spline basis functions.
}
\examples{
# Example using artificially generated data
set.seed(1)
n <- 40
x1 <- runif(n)
x2 <- runif(n)
x <- cbind(x1,x2)
F1 <- function(x1,x2) {
  (x1 - 2*x2)^2/sqrt(x1^2 + 4*x2^2)
}
F2 <- function(x1,x2) {
  (3*x1 + x2)^2/sqrt(9*x1^2 + x2^2)
}
plotGrid(list(window = list(x = range(x1), y = range(x2)),
              x = x, def.x = cbind(F1(x1,x2), F2(x1,x2))),
         F1 = F1, F2 = F2, margins = TRUE)
covModel <- RMmatern(nu = 2.5, var = 1, scale = 2) + RMnugget(var = 1)
data <- RFsimulate(covModel, x = F1(x1,x2), y = F2(x1,x2))
y <- as.numeric(unlist(data@data))
test1 <- bdef(x, y) # lambda estimated
test2 <- bdef(x, y, lambda = .5) # lambda provided by user

}
\author{
Guilherme Ludwig and Ronaldo Dias
}
\references{
To add.
}
\seealso{
\code{\link{RandomFields}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

