% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGDdist.R
\name{plotGDdist}
\alias{plotGDdist}
\title{Scatterplot of sample spatial covariance entries and pairwise distances in G- and D-domain}
\usage{
plotGDdist(model, plot = TRUE, lines = TRUE)
}
\description{
This function produces a plot of the upper diagonal sample covariance entries and the
corresponding pairwise distances in the original and deformed coordinates, where the
spatial deformation is obtained with a tensor product of B-spline basis functions,
implemented in the \code{\link{bdef}} function.
}

