% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Plot Spatial Deformations from bdef}
\usage{
plotGrid(model, nx = 20, ny = 20, colorO = "Black", colorD = "Red",
  plot = TRUE, margins = FALSE, persp = FALSE, F1 = NULL, F2 = NULL)
}
\arguments{
\item{model}{An object returned by the function \code{\link{bdef}}.
Alternatively, a list with named components matching
those returned by \code{\link{bdef}}.}

\item{nx}{Number of ticks in x-axis for grid. Defaults to 10.}

\item{ny}{Number of ticks in y-axis for grid. Defaults to 10.}

\item{colorO}{Color of original data points and grid. Defaults to Black.}

\item{colorD}{Color of deformed data points and grid. Defaults to Red.}

\item{plot}{Whether a plot should be produced or not. Defaults to TRUE.}

\item{margins}{Whether a layout should be produced, with f1, f2 estimates
shown along the corresponding axes. Defaults to FALSE.}

\item{F1}{Allows the user to supply a function that generates the deformed
coordinate y1. Defaults to NULL, in which case the
tensor product of B-splines is used instead.}

\item{F2}{Allows the user to supply a function that generates the deformed
coordinate y2. Defaults to NULL, in which case the
tensor product of B-splines is used instead.}
}
\value{
\code{plotGrid} invisibly returns the values of the deformation on
                    the grid.
}
\description{
This function plot a spatial deformation on a grid obtained via the bdef
function
}
\examples{
# Example using artificially generated data
set.seed(1)
n <- 10
x1 <- runif(n)
x2 <- runif(n)
x <- cbind(x1,x2)
\dontrun{
   plot(x, xlim = c(0,1), ylim = c(0,1), pch = letters[1:n])
}
K <- 8
theta1 <- exp(-0.5*scale(kronecker(1:K - K/2,
              rep(1,K) + 5*((1:K - K/2)^2)/K^2),
              center = FALSE, scale = TRUE))
theta2 <- scale(kronecker(rep(1,K) - ((1:K - K/2)^2)/K^2 + rnorm(K, 0, .1),
          1:K - K/2),
          center = FALSE, scale = TRUE)
tempb1 <- bs(x[, 1], K, intercept = TRUE)
tempb2 <- bs(x[, 2], K, intercept = TRUE)
tempW <- matrix(0, n, K^2)
for(i in 1:n) tempW[i,] <- kronecker(tempb1[i,], tempb2[i,])
dx <- cbind(tempW\%*\%theta1, tempW\%*\%theta2)
fakeModel <- list(basis = list(B1 = bs(range(x[, 1]), K, intercept = TRUE),
                               B2 = bs(range(x[, 2]), K, intercept = TRUE)),
                  window = list(x = 0:1, y = 0:1),
                  x = x, dx = dx,
                  theta1 = theta1,
                  theta2 = theta2,
                  df1 = K,
                  df2 = K)
 plotGrid(fakeModel)
 plotGrid(fakeModel, margins = TRUE)
 plotGrid(fakeModel, persp = TRUE)
 # pdf("sim1.pdf"); plotGrid(fakeModel); dev.off()
 # pdf("sim2.pdf"); plotGrid(fakeModel, margins = TRUE); dev.off()
 # pdf("sim3.pdf"); plotGrid(fakeModel, persp = TRUE); dev.off()

}
\author{
Guilherme Ludwig and Ronaldo Dias
}
\references{
To add.
}
\seealso{
\code{\link{geoR::likfit}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

