% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdef.R
\name{bdef}
\alias{bdef}
\title{Spatial Deformation via tensor product of B-splines}
\usage{
bdef(x, y, tim = NULL, cov.model = RMexp(var = NA, scale = NA) +
  RMnugget(var = NA), df1 = 6, df2 = 6, window = list(x = range(x[, 1]), y
  = range(x[, 2])), lambda = NA, maxit = 2, traceback = FALSE,
  fullDes = TRUE, ...)
}
\arguments{
\item{x}{A n times 2 matrix of coordinates for the sampled data points. The
rows must be unique, corresponding to distinct sample
locations.}

\item{y}{A vector of length n*m with the values taken by the response
at the corresponding spatial locations in x, for
each time point t_1, ..., t_m.}

\item{tim}{A vector of equally spaced time positions, or NULL, in which
the temporal component of the model is ignored.
Defaults to NULL. Each time point must have the
same corresponding locations as rows in x.}

\item{cov.model}{A model for the spatial covariance function. See
\code{\link{RMmodel}} from the RandomFields package
for details. Defaults to Exponential model with
unknown variance and scale, plus a nugget effect. Right
now the model doesn't allow space-time covariance models.}

\item{df1}{Degrees of freedom for the tensor product of splines along
the x[,1] coordinate. Defaults to 6.}

\item{df2}{Degrees of freedom for the tensor product of splines along
the x[,2] coordinate. Defaults to 6.}

\item{window}{A list with two named entries, "x" and "y", each a length 2
numeric vector with the minimum and maximum values
taken for the respective coordinate in the sampled
region. Defaults to the range of the columns in
argument x.}

\item{lambda}{Penalization parameter. If set to NA (default), this
parameter will be estimated. Otherwise, it will use
the user supplied value. The parameter is in a log
scale, thus ranging from -Inf to +Inf.}

\item{maxit}{The maximum number of iterations between
optimization of the covariance parameters
and estimation of the spline coefficients.
Defaults to 2.}

\item{traceback}{Whether the function will return each step
of the optimization iterations, as a list.
For debugging purposes. Defaults to FALSE.}

\item{fullDes}{Whether the returned object will include a processed
copy of the dataset. Needed for the
\code{\link{plotGDdist}} function, but can be set
to FALSE if the user has no interest in GDdist.
Defaults to TRUE.}

\item{...}{Additional arguments for RFfit.}
}
\value{
\code{bdef} returns an object of \code{\link{class}} "bdef"
containing at least the following components
  \item{window}{The rectangular spatial domain in which the
               data was sampled}
  \item{basis}{list of B-spline basis functions, inherited from
               \code{\link{bs}}, with one set for each coordinate}
  \item{theta1}{B-spline coefficients for the coordinate x1}
  \item{theta2}{B-spline coefficients for the coordinate x2}
}
\description{
This function finds a spatial deformation for Gaussian processes, based
on the penalized log-likelihood function. The deformation is obtained
via the tensor product of B-spline basis functions.
}
\examples{
# Example using artificially generated data
set.seed(1)
n <- 40
x1 <- runif(n)
x2 <- runif(n)
x <- cbind(x1,x2)
F1 <- function(x1,x2) {
  x1 # (x1 - 2*x2)^2/sqrt(x1^2 + 4*x2^2)
}
F2 <- function(x1,x2) {
  sqrt(x1*x2) # (3*x1 + x2)^2/sqrt(9*x1^2 + x2^2)
}
plotGrid(list(window = list(x = range(x1), y = range(x2)),
              x = x, def.x = cbind(F1(x1,x2), F2(x1,x2))),
         F1 = F1, F2 = F2, margins = TRUE)
covModel <- RMexp(var = 1, scale = 2) + RMnugget(var = 1)
data <- RFsimulate(covModel, x = F1(x1,x2), y = F2(x1,x2))
y <- as.numeric(unlist(data@data))
test1 <- bdef(x, y) # lambda estimated
plotGrid(test1, margins = TRUE)
\dontrun{
test2 <- bdef(x, y, lambda = -100) # lambda provided by user
test3 <- bdef(x, y, lambda = 100) # ditto
test4 <- bdef(x, y, cov.model = RMmatern(nu = 2.5, var = NA, scale = NA)) # No nugget effect
plotGrid(test2, margins = TRUE)
plotGrid(test3, margins = TRUE)
plotGrid(test4, margins = TRUE)
}
# Time version
TIME <- 20
covModel2 <- RMexp(proj = "space", var = 1, scale = 1) *
  RMnugget(proj = "time", var = 1)
data2 <- RFsimulate(covModel2, x = F1(x1,x2), y = F2(x1,x2), T = 1:TIME)
y <- as.numeric(unlist(data2@data))
\\notrun{
matplot(1:TIME, t(matrix(y, ncol = TIME)), type = "b", lty = 1, pch = 1)
}
covModel2m <- RMexp(var = NA, scale = NA) # Only space?
testT <- bdef(x, y, tim = 1:TIME, cov.model = covModel2m) # lambda estimated
plotGrid(testT, margins = TRUE)

}
\author{
Guilherme Ludwig and Ronaldo Dias
}
\references{
To add.
}
\seealso{
\code{\link{RandomFields}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

